Создал матрицу трансформаций для координат. Теперь объекты лежат в [-1;1] и на выходе преобразуются в координаты по пикселям. Добавлены адаптивные карты. В связи с тем, что все до этого было завязано на координатах, было переделано много логики. Появился отдельный класс для коллизий, который с хорошей точностью отслеживает столкновения. Есть возможность однопользовательской и 2 пользовательской игры(включается пока только хардово, через коммент создания нового user-танка). Созданы глупые боты, который пока что ездят все синхроно в одном направлении. Нужно сделать: боты(больше рандома), убирать убитые объекты(чтобы не стреляли после смерти), сцены для выбора режима и кол-во игроков, по хорошему счет в битвах  механизм продолжения игры без перезахода в играть. В идеале сделать парсер для создания карт из файликов.

+++ src/main/java/com/ru/vsu/cs/dplatov/vvp/tanks2d/objects/GameObject.java
@@ -3,28 +3,28 @@ package com.ru.vsu.cs.dplatov.vvp.tanks2d.objects;
 import javafx.scene.image.ImageView;
 
 abstract public class GameObject {
-    private int x, y;
+    private double x, y;
     private ImageView view;
 
-    public GameObject(int x, int y, ImageView view) {
+    public GameObject(double x, double y, ImageView view) {
         this.x = x;
         this.y = y;
         this.view = view;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
